refactor: rename common type files
* it needs to be decent `.ts` file to create declaration files
