+++ apps/calendar/src/jest.d.ts
@@ -10,5 +10,3 @@ declare global {
     }
   }
 }
-
-export {};
