--- apps/calendar/types/toolbar.d.ts
@@ -1,7 +1,0 @@
-import type { ViewListMap } from '@t/options';
-
-interface ToolbarProps {
-  viewName: string;
-  components: ViewListMap;
-  go: (viewName: string) => void;
-}
