--- apps/calendar/types/util.d.ts
@@ -1,6 +1,0 @@
-interface Element {
-  msMatchesSelector: (selectors: string) => boolean;
-}
-
-type MouseEventListener = (e: MouseEvent) => void;
-type KeyboardEventListener = (e: KeyboardEvent) => void;
